+++ changelog.md
@@ -30,7 +30,7 @@
   added support for emulators with string skin names
 - Fixed loading of keystores with no password
 - Send connection: close header in responses. Fixes [#458](../../issues/458)
-
+- Removed 'SelendroidKeys.ANDROID_HOME' because emulating the home key using Instrumentation is not supported in Android. Use adb to emulate the home key in your tests.
 
 0.11.0
 ---